requirements.txt
amqp==5.0.6
asgiref==3.4.1
beautifulsoup4==4.10.0
billiard==3.6.4.0
bs4==0.0.1
celery==5.1.2
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.6
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==35.0.0
defusedxml==0.7.1
Django==3.2.7
django-allauth==0.45.0
django-ckeditor==6.1.0
django-debug-toolbar==3.2.2
django-js-asset==1.2.2
django-modeltranslation==0.17.3
django-recaptcha3==0.4.0
flake8==4.0.1
idna==3.2
kombu==5.1.0
mccabe==0.6.1
oauthlib==3.1.1
Pillow==8.3.2
prompt-toolkit==3.0.20
pycodestyle==2.8.0
pycparser==2.20
pyflakes==2.4.0
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.1
redis==3.5.3
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
flake8-blind-except==0.2.0
flake8-builtins==1.5.3
flake8-cognitive-complexity==0.1.0
flake8-commas==2.0.0
flake8-comprehensions==3.6.1
flake8-django==1.1.2
flake8-eradicate==1.1.0
flake8-expression-complexity==0.0.9
flake8-logging-format==0.6.0
flake8-polyfill==1.0.2
flake8-print==4.0.0
flake8-variables-names==0.0.4
pep8-naming